chore: rename canyon_events.h to moth_graphics_events.h and minor housekeeping
- Rename events/canyon_events.h to events/moth_graphics_events.h
- Rename SpriteSheet::clips member to m_clips for convention consistency
- Replace stale 'canyon' references with 'moth_graphics' in doc comments
